Join classes in unique file
