+++ cpp/src/gandiva/CMakeLists.txt
@@ -92,7 +92,6 @@ set(SRC_FILES
     replace_holder.cc
     selection_vector.cc
     tree_expr_builder.cc
-    to_date_holder.cc
     random_generator_holder.cc
     ${GANDIVA_PRECOMPILED_CC_PATH})
 
@@ -235,7 +234,6 @@ add_gandiva_test(internals-test
                  greedy_dual_size_cache_test.cc
                  lru_cache_test.cc
                  to_date_holder_test.cc
-                 to_time_holder.cc
                  to_time_holder_test.cc
                  simple_arena_test.cc
                  like_holder_test.cc
