+++ cpp/src/gandiva/function_holder_registry.h
@@ -28,8 +28,8 @@
 #include "gandiva/like_holder.h"
 #include "gandiva/node.h"
 #include "gandiva/random_generator_holder.h"
+#include "gandiva/to_date_functions_holder.h"
 #include "gandiva/replace_holder.h"
-#include "gandiva/to_date_holder.h"
 #include "gandiva/to_time_holder.h"
 
 namespace gandiva {
