+++ cpp/src/gandiva/gdv_function_stubs.cc
@@ -41,9 +41,8 @@
 #include "gandiva/like_holder.h"
 #include "gandiva/precompiled/types.h"
 #include "gandiva/random_generator_holder.h"
+#include "gandiva/to_date_functions_holder.h"
 #include "gandiva/replace_holder.h"
-#include "gandiva/to_date_holder.h"
-#include "gandiva/to_time_holder.h"
 
 /// Stub functions that can be accessed from LLVM or the pre-compiled library.
