+++ cpp/src/gandiva/to_date_holder_test.cc
@@ -21,7 +21,7 @@
 #include "arrow/testing/gtest_util.h"
 
 #include "gandiva/execution_context.h"
-#include "gandiva/to_date_holder.h"
+#include "gandiva/to_date_functions_holder.h"
 #include "precompiled/epoch_time_point.h"
 
 #include <gtest/gtest.h>
